+++ src/transformers/modeling_bart.py
@@ -51,15 +51,8 @@ _CONFIG_FOR_DOC = "BartConfig"
 _TOKENIZER_FOR_DOC = "BartTokenizer"
 
 
-BART_PRETRAINED_MODEL_ARCHIVE_LIST = [
-    "facebook/bart-base",
-    "facebook/bart-large",
-    "facebook/bart-large-mnli",
-    "facebook/bart-large-cnn",
-    "facebook/bart-large-xsum",
-    "facebook/mbart-large-en-ro",
-    # See all BART models at https://huggingface.co/models?filter=bart
-]
+# See all BART models at https://huggingface.co/models?filter=bart
+
 
 
 BART_START_DOCSTRING = r"""
