+++ src/transformers/tokenization_pegasus.py
@@ -133,7 +133,8 @@ class PegasusTokenizer(ReformerTokenizer):
             return model_inputs
         if max_target_length is not None:
             tokenizer_kwargs["max_length"] = max_target_length
-        decoder_inputs: BatchEncoding = self(tgt_texts, **tokenizer_kwargs)
-        for k, v in decoder_inputs.items():
-            model_inputs[f"decoder_{k}"] = v
+        labels: BatchEncoding = self(tgt_texts, **tokenizer_kwargs)['input_ids']
+        model_inputs['labels'] = labels
+        #for k, v in decoder_inputs.items():
+        #    model_inputs[f"decoder_{k}"] = v
         return model_inputs
